+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Method for deleting a node recursively
+
+### Fixed 
+- Fixed `Group::deleteLink` and `DataSet::deleteLink` for nested nodes - #48
+
 ## [2.2.3] - 2023-01-13
 ### Fixed 
 - Fixed a bug in `Data::resize` method returning false negatives - #47